separate scoreboard into presentational and container
