+++ app/main.jsx
@@ -3,7 +3,7 @@ import React from 'react'
 import {Router, Route, IndexRedirect, browserHistory} from 'react-router'
 import {render} from 'react-dom'
 
-import Scoreboard from './components/Scoreboard'
+import Scoreboard from './components/container/ScoreboardContainer'
 import ActiveSpymasterHint from './components/ActiveSpymasterHint'
 import 'semantic-ui-css/semantic.min.css'
